SonicClient.TcpConnection (sonic_client v0.1.0)

This is the TcpConnection module, responsible to send and receive calls.

Functions

close(conn)

open(host, port, opts \\ [], timeout \\ 5000)

Specs

open(String.t(), integer(), list(), integer()) :: {:error, any()} | {:ok, pid()}

Starts connection with the Sonic server as a child process.

Returns {:ok, conn}.

Where conn is the PID of the client.

Examples

SonicClient.TcpConnection.open(127.0.0.1, 1491, [])
{:ok, #PID<0.198.0>}

request(conn, command)

Specs

request(pid(), String.t()) :: {:ok, String.t()} | {:error, any()}

Sends a synchronous request with command command to the Sonic server through the PID child.

Returns {:ok, response}.

Where response is a String with the body of the response.